services/user_service/requirements/lock.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
blinker==1.9.0
boto3==1.39.5
botocore==1.39.5
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
Deprecated==1.2.18
fastapi==0.116.1
Flask==3.1.1
googleapis-common-protos==1.70.0
grpcio==1.73.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-instrumentation==0.56b0
opentelemetry-instrumentation-asgi==0.56b0
opentelemetry-instrumentation-fastapi==0.56b0
opentelemetry-instrumentation-flask==0.56b0
opentelemetry-instrumentation-requests==0.56b0
opentelemetry-instrumentation-wsgi==0.56b0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opentelemetry-util-http==0.56b0
packaging==25.0
protobuf==6.31.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
requests==2.32.4
s3transfer==0.13.0
setuptools==80.9.0
-e git+ssh://git@github.com/mcdaigle1/gsalegig_utils.git@9818033801a36e618b4c5577bb99f0ac9f9bdd20#egg=shared_utils
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
thrift==0.22.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
